backend/requirements.txt
flask
flask-cors
openai
requests
python-dotenv
PyJWT
replicate